1609 : changes based on comments
