+++ templates/front/nouveau_formulaire.html.twig
@@ -19,7 +19,6 @@
            data-ajaxurl-get-signalement-draft="{{ path('informations_signalement_draft', {uuid: uuid_signalement}) }}"
          {% endif %}
          data-platform-name="{{ platform.name }}"
-         data-platform-url="{{ platform.url }}"
         ></div>
   </div>
 {% endblock %}
