+++ tools/wiremock/src/Resources/Signalement/questions_profile_bailleur.json
@@ -1165,7 +1165,7 @@
           "label": "Retour à l'accueil",
           "slug": "validation_signalement_homepage",
           "customCss": "fr-btn",
-          "link": "{{formStore.props.platformUrl}}"
+          "link": "/"
         }
       ]
     }
