+++ tools/wiremock/src/Resources/Signalement/questions_profile_service_secours.json
@@ -799,7 +799,7 @@
           "label": "Retour à l'accueil",
           "slug": "validation_signalement_homepage",
           "customCss": "fr-btn",
-          "link": "{{formStore.props.platformUrl}}"
+          "link": "/"
         }
       ]
     }
