+++ tools/wiremock/src/Resources/Signalement/questions_profile_tiers_particulier.json
@@ -1040,7 +1040,7 @@
           "label": "Retour à l'accueil",
           "slug": "validation_signalement_homepage",
           "customCss": "fr-btn",
-          "link": "{{formStore.props.platformUrl}}"
+          "link": "/"
         }
       ]
     }
